// src/shared/api/schema.ts
export interface components {
    schemas: {
// ...
        UserBaseSchema: {
            /** Id */
            id: number;
            /** First Name */
            first_name?: string | null;
            /** Last Name */
            last_name?: string | null;
            /** Patronymic */
            patronymic?: string | null;
            /** Phone */
            phone: string;
            /** Email */
            email?: string | null;
            /**
             * Join Date
             * Format: date-time
             */
            join_date: string;
            /** Birth Day */
            birth_day?: string | null;
            /** Comment */
            comment?: string | null;
            role: components["schemas"]["EnumUserRoles"];
            type: components["schemas"]["EnumUserTypes"];
            /** Company Type */
            company_type?: string | null;
            /** Company Name */
            company_name?: string | null;
            /** Company Number */
            company_number?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}